cephadm: move registry_login to container_engines.py
